+++ plugins/postcss-fit-tolerance-property/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Fit Tolerance Property
 
+### Unreleased (patch)
+
+- Skip properties with a fallback
+
 ### 1.0.0
 
 _August 15, 2026_