+++ plugins/postcss-private-rule/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Private Rule
 
+### Unreleased (patch)
+
+- Only assign new property names to the PostCSS AST when the name actually changed
+
 ### 1.0.0
 
 _August 30, 2026_